// types/graphql-types.d.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  /** The `ID` scalar type represents a unique identifier, often used to refetch an object or as key for a cache. The ID type appears in a JSON response as a String; however, it is not intended to be human-readable. When expected as an input type, any string (such as `"4"`) or integer (such as `4`) input value will be accepted as an ID. */
  ID: string;
  /** The `String` scalar type represents textual data, represented as UTF-8 character sequences. The String type is most often used by GraphQL to represent free-form human-readable text. */
  String: string;
  /** The `Boolean` scalar type represents `true` or `false`. */
  Boolean: boolean;
  /** The `Int` scalar type represents non-fractional signed whole numeric values. Int can represent values between -(2^31) and 2^31 - 1. */
  Int: number;
  /** The `Float` scalar type represents signed double-precision fractional values as specified by [IEEE 754](https://en.wikipedia.org/wiki/IEEE_floating_point). */
  Float: number;
  /** A date string, such as 2007-12-03, compliant with the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  Date: any;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
  /**
   * The `DateTime` scalar type represents a DateTime
   * value as specified by
   * [iso8601](https://en.wikipedia.org/wiki/ISO_8601).
   */
  Portfolio_DateTime: any;
};
// ...
export type SitePageContext = {
  category?: Maybe<Scalars['String']>;
  prefix?: Maybe<Scalars['String']>;
  currentPage?: Maybe<Scalars['Int']>;
  numberOfPages?: Maybe<Scalars['Int']>;
  first?: Maybe<Scalars['Int']>;
  after?: Maybe<Scalars['String']>;
  hasNextPage?: Maybe<Scalars['Boolean']>;
  hasPreviousPage?: Maybe<Scalars['Boolean']>;
  title?: Maybe<Scalars['String']>;
  categories?: Maybe<Array<Maybe<Scalars['String']>>>;
  slug?: Maybe<Scalars['String']>;
  previousPath?: Maybe<Scalars['String']>;
  nextPath?: Maybe<Scalars['String']>;
};
// ...
export type Portfolio_BlogPostType = Portfolio_Node & {
  slug: Scalars['String'];
  title: Scalars['String'];
  introduction: Scalars['String'];
  markdown: Scalars['String'];
  publishedAt?: Maybe<Scalars['Portfolio_DateTime']>;
  updatedAt: Scalars['Portfolio_DateTime'];
  /** The ID of the object. */
  id: Scalars['ID'];
  image?: Maybe<Portfolio_ImageType>;
  category?: Maybe<Scalars['String']>;
};

/** An object with an ID */
export type Portfolio_Node = {
  /** The ID of the object. */
  id: Scalars['ID'];
};


export type Portfolio_ImageType = {
  title: Scalars['String'];
  url?: Maybe<Scalars['String']>;
};
// ...
export type GetBlogPostsQuery = { portfolio: { blogPosts?: Maybe<{ edges: Array<Maybe<{ node?: Maybe<(
          Pick<Portfolio_BlogPostType, 'slug' | 'title' | 'introduction' | 'publishedAt' | 'category'>
          & { image?: Maybe<Pick<Portfolio_ImageType, 'url' | 'title'>> }
        )> }>> }> } };